app/requirements.txt
Flask
cassandra-driver
Faker